# server/config/seed_data/swagger.json
{
  "swagger": "2.0",
  "paths": {
...
    "/api/organizations/{id}/users/{userId}": {
      "get": {
          "summary": "/api/organizations/{id}/users/{userId}",
          "consumes": [
              "application/json"
          ],
          "parameters": [
              {
                  "name": "id",
                  "in": "path",
                  "required": true
              },
              {
                  "name": "userId",
                  "in": "path",
                  "required": true
              }
          ],
          "responses": {}
      }
    },
...
    "/api/users/{id}": {
      "delete": {
        "summary": "Delete user by id",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "User Identifier"
          }
        ],
        "responses": {},
        "tags": [
          "users"
        ],
        "security": {
          "Bearer": []
        }
      },
      "put": {
        "summary": "Update user by id",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "User Identifier"
          },
          {
            "consumes": [
              "application/json"
            ],
            "in": "body",
            "description": ""
          }
        ],
        "responses": {},
        "tags": [
          "users"
        ],
        "security": {
          "Bearer": []
        }
      },
      "get": {
        "summary": "Get user by id",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "User Identifier"
          }
        ],
        "responses": {},
        "tags": [
          "users"
        ],
        "security": {
          "Bearer": []
        }
      }
    },
...
    "/api/users/": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get users",
        "security": {
          "Bearer": []
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create user",
        "parameters": [
          {
            "consumes": [
              "application/json"
            ],
            "name": "body",
            "in": "body",
            "description": ""
          }
        ]
      }
    }
  },
  "info": {
    "title": "Health Equity Colorado API Documentation",
    "version": "1.0.0",
    "description": "This is the documentation for the Health Equity Colorado website. You can find the production site at [https://health-equity-colorado-prod.herokuapp.com/](https://health-equity-colorado-prod.herokuapp.com/)."
  },
...
}